Align stories to .ts files. Align grid layout to support the following structure: (string | number)[][].
